Fixed parse_scandir loop malfunction.
parse_scandir is only dispatching work to parse_plist by looping
on the given directory.

The code must be a remainder of a previous Class implementation,
because it was trying to iterate on an Analysis object (result of
parse_plist) that is non-iterable.

Instead, yield the Analysis per plist report until they are processed.